Refactor Searchbar into Smaller Components (#237)
* Split large searchbar component into proper hierarchy

* Clean up

* Clean up icon button and search text
